boarding_house/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
psycopg2-binary
daphne
